refactor: share research fetch helpers
