+++ web/src/components/research/ResearchDashboard.tsx
@@ -2,6 +2,7 @@
 
 import { useEffect, useMemo, useState } from "react"
 
+import { fetchJson, getErrorMessage } from "@/lib/fetch"
 import { Badge } from "@/components/ui/badge"
 import { Button } from "@/components/ui/button"
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card"
@@ -89,52 +90,6 @@ type ConfirmAction =
   | { type: "bulk_moderate"; status: "approved" | "rejected"; itemIds: number[] }
   | { type: "bulk_move"; itemIds: number[]; targetTrackId: number }
   | { type: "clear_track_memory"; trackId: number }
-
-type UpstreamErrorBody = {
-  detail?: string
-  error?: string
-}
-
-function toFriendlyErrorMessage(status: number, rawText: string): string | null {
-  if (!rawText) return null
-
-  let parsed: UpstreamErrorBody | null = null
-  try {
-    parsed = JSON.parse(rawText) as UpstreamErrorBody
-  } catch {
-    parsed = null
-  }
-
-  const detail = parsed && typeof parsed.detail === "string" ? parsed.detail : undefined
-
-  if (detail && (detail.includes("Upstream API unreachable") || detail.includes("Upstream API timed out"))) {
-    if (detail.includes("timed out")) {
-      return "Unable to connect to service (request timed out). Please ensure the backend is running. Please try again."
-    }
-    return "Unable to connect to service. Please ensure the backend is running."
-  }
-
-  // Fallback: surface backend-provided detail when available
-  if (detail) return detail
-
-  return null
-}
-
-async function fetchJson<T>(url: string, init?: RequestInit): Promise<T> {
-  const res = await fetch(url, init)
-  if (!res.ok) {
-    const text = await res.text().catch(() => "")
-    const friendly = toFriendlyErrorMessage(res.status, text)
-    if (friendly) {
-      throw new Error(friendly)
-    }
-    const statusLabel = `${res.status} ${res.statusText}`.trim()
-    const base = statusLabel || "Request failed"
-    const message = text ? `${base} ${text}`.trim() : base
-    throw new Error(message)
-  }
-  return res.json() as Promise<T>
-}
 
 function clampNumber(value: number, min: number, max: number, fallback: number) {
   if (!Number.isFinite(value)) return fallback
@@ -256,7 +211,7 @@ export default function ResearchDashboard() {
 
   useEffect(() => {
     setError(null)
-    refreshTracks().catch((e) => setError(e instanceof Error ? e.message : String(e)))
+    refreshTracks().catch((e) => setError(getErrorMessage(e)))
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [])
 
@@ -309,7 +264,7 @@ export default function ResearchDashboard() {
         await refreshTracks()
       }
     } catch (e) {
-      setError(e instanceof Error ? e.message : String(e))
+      setError(getErrorMessage(e))
     } finally {
       setLoading(false)
     }
@@ -334,7 +289,7 @@ export default function ResearchDashboard() {
       setSuggestText("")
       await refreshInbox(activeTrackId)
     } catch (e) {
-      setError(e instanceof Error ? e.message : String(e))
+      setError(getErrorMessage(e))
     } finally {
       setLoading(false)
     }
@@ -357,7 +312,7 @@ export default function ResearchDashboard() {
       })
       await refreshInbox(activeTrackId)
     } catch (e) {
-      setError(e instanceof Error ? e.message : String(e))
+      setError(getErrorMessage(e))
     } finally {
       setLoading(false)
     }
@@ -381,7 +336,7 @@ export default function ResearchDashboard() {
       })
       await refreshInbox(activeTrackId)
     } catch (e) {
-      setError(e instanceof Error ? e.message : String(e))
+      setError(getErrorMessage(e))
     } finally {
       setLoading(false)
     }
@@ -420,7 +375,7 @@ export default function ResearchDashboard() {
       const activeId = await refreshTracks()
       await refreshInbox(activeId)
     } catch (e) {
-      setError(e instanceof Error ? e.message : String(e))
+      setError(getErrorMessage(e))
     } finally {
       setLoading(false)
     }
@@ -458,7 +413,7 @@ export default function ResearchDashboard() {
       })
       await buildContext(false)
     } catch (e) {
-      setError(e instanceof Error ? e.message : String(e))
+      setError(getErrorMessage(e))
     } finally {
       setLoading(false)
     }
@@ -474,7 +429,7 @@ export default function ResearchDashboard() {
       )
       await refreshInbox(trackId)
     } catch (e) {
-      setError(e instanceof Error ? e.message : String(e))
+      setError(getErrorMessage(e))
     } finally {
       setLoading(false)
     }
@@ -973,7 +928,7 @@ export default function ResearchDashboard() {
                       />
                       <Button
                         variant="outline"
-                        onClick={() => refreshEval().catch((e) => setError(e instanceof Error ? e.message : String(e)))}
+                        onClick={() => refreshEval().catch((e) => setError(getErrorMessage(e)))}
                         disabled={loading}
                       >
                         Refresh
@@ -989,7 +944,7 @@ export default function ResearchDashboard() {
                   </div>
 
                   {!evalSummary ? (
-                    <Button variant="secondary" onClick={() => refreshEval().catch((e) => setError(e instanceof Error ? e.message : String(e)))} disabled={loading}>
+                    <Button variant="secondary" onClick={() => refreshEval().catch((e) => setError(getErrorMessage(e)))} disabled={loading}>
                       Load Summary
                     </Button>
                   ) : (
